PVS V1028: cast result of operation instead of operands
If an overflow occurs before the cast, we can't fix it. If we cast
first, we can rely on integer promotion to make the result use the
appropriate size.

Reviewed-on: https://review.haiku-os.org/c/1623

+++ src/add-ons/media/media-add-ons/usb_webcam/addons/sonix/SonixCamDevice.cpp
@@ -634,10 +634,10 @@ SonixCamDevice::MinRawFrameSize()
 {
 	// if (fCompressionEnabled) { ... return ; }
 	BRect vf(VideoFrame());
-	int w = vf.IntegerWidth()+1;
-	int h = vf.IntegerHeight()+1;
+	size_t w = vf.IntegerWidth()+1;
+	size_t h = vf.IntegerHeight()+1;
 	// 1 byte/pixel
-	return (size_t)(w*h);
+	return w * h;
 }
